⚡ Add parallax and remove js for sparks
From README of https://github.com/cristurm/nyan-cat who want to get rid of JS
for sparks and add parallax.
As I did a small test for my personal project, I added parallax and get rid of
JS to replicate sparks using nth-child and flex box.

The downside of this commit is that replication of sparkle are null as we need
to put sparkle div manually. But we can but more sparkle than need as
animation are conditionnal to the nth-child.